+++ src/views/components/molecules/Ditto/index.tsx
@@ -1,88 +1,65 @@
-import React, { useEffect } from 'react';
-import styled from 'styled-components';
+import React, { useEffect, useRef } from 'react';
 import moment from 'moment';
-import { transparentize } from 'polished';
 import { TimelineLite, Power1 } from 'gsap';
-import parse from 'html-react-parser';
-import './index.style.scss';
-
-interface StyledProps {
-  imgSrc?: string;
-}
-const DittoRoot = styled.div``;
-const Main = styled.div``;
-const Body = styled.div``;
-const Title = styled.div``;
-const Footer = styled.div``;
-const Thumnail = styled.div<StyledProps>`
-  background-image: url(${({ imgSrc }) => imgSrc});
-  background-color: ${({ theme }) =>
-    transparentize(0.75, theme.color.grayscales.dark[4])};
-`;
-const MainInner = styled.div`
-  background-color: ${({ theme }) => theme.color.bg[0]};
-`;
-const Content = styled.div`
-  color: ${({ theme }) => theme.color.text[2]};
-`;
-const Date = styled.div`
-  color: ${({ theme }) => theme.color.text[2]};
-`;
-const CategoryLabel = styled.div`
-  background-color: ${({ theme }) => theme.color.bg[1]};
-`;
+import {
+  DittoRoot,
+  Main,
+  Body,
+  Title,
+  Footer,
+  Thumnail,
+  MainInner,
+  Content,
+  Date,
+  CategoryLabel,
+} from './index.style';
 
 interface DittoProps {
-  excerpt: string;
-  date: number;
-  footer: () => React.ReactNode;
-  imgSrc?: string;
-  title: () => React.ReactNode;
-  isGrid: boolean;
+  data: {
+    excerpt: string;
+    date: number;
+    footer: () => React.ReactNode;
+    imgSrc?: string;
+    title: () => React.ReactNode;
+    isGrid: boolean;
+  };
 }
-const Ditto = ({
-  imgSrc,
-  excerpt,
-  date,
-  footer,
-  title,
-  isGrid,
-}: DittoProps) => {
+const Ditto = ({ data }: DittoProps) => {
+  const { imgSrc, excerpt, date, footer, title, isGrid } = data;
+  const MainInnerRef = useRef<HTMLDivElement>(null);
   const _excerpt = () => {
-    if (excerpt.length < 85) return excerpt;
-    return `${excerpt.substring(0, 80)}...`;
+    if (excerpt.length < 118) return excerpt;
+    return `${excerpt.substring(0, 110)}...`;
   };
   useEffect(() => {
     const tl = new TimelineLite();
     if (isGrid) {
-      tl.to(['._mainInner'], 0.05, {
+      tl.to(MainInnerRef.current, 0.05, {
         marginLeft: '0',
-      }).to(['._mainInner'], 0.2, {
+      }).to(MainInnerRef.current, 0.2, {
         marginTop: '54%',
         ease: Power1.easeOut,
       });
     } else {
-      tl.to(['._mainInner'], 0.25, {
+      tl.to(MainInnerRef.current, 0.25, {
         style: null,
       });
     }
   });
   return (
-    <DittoRoot
-      className={`ditto ditto-root ditto-${isGrid ? 'grid' : 'list'} ${
-        imgSrc ? '' : 'non-thumbnail'
-      }`}
-    >
-      <Main className="_main">
-        <Thumnail className="_thumnail" imgSrc={imgSrc} />
-        <MainInner className="_mainInner">
-          <Body className="_body">
-            <Title className="_title">{title()}</Title>
-            <Content className="_content">{parse(_excerpt())}</Content>
+    <DittoRoot>
+      <Main isGrid={isGrid} isThumnail={!!imgSrc}>
+        <Thumnail imgSrc={imgSrc} isGrid={isGrid} isThumnail={!!imgSrc} />
+        <MainInner ref={MainInnerRef} isGrid={isGrid} isThumnail={!!imgSrc}>
+          <Body isGrid={isGrid} isThumnail={!!imgSrc}>
+            <Title>{title()}</Title>
+            <Content isThumnail={!!imgSrc} isGrid={isGrid}>
+              {_excerpt()}
+            </Content>
           </Body>
-          <Footer className="_footer">
-            <CategoryLabel className="_cat">{footer()}</CategoryLabel>
-            <Date className="_date">{moment(date).format('YYYY.M.D')}</Date>
+          <Footer>
+            <CategoryLabel>{footer()}</CategoryLabel>
+            <Date>{moment(date).format('YYYY.M.D')}</Date>
           </Footer>
         </MainInner>
       </Main>