+++ src/views/components/molecules/SyntaxHighlighter/index.tsx
@@ -39,7 +39,7 @@ const SyntaxHighlighter: React.FC<Props> = (props: Props) => {
         {fileName ? <FileName>{fileName}</FileName> : <div />}
         {lang && (
           <Lang
-            src={`https://wp.hapas.io/wp-content/uploads/icons/material/${icon}.svg`}
+            src={`https://wp.hapas.io/wp-content/uploads/4log/icons/material/${icon}.svg`}
             alt=""
           />
         )}